〇

+++ src/main/java/scoremanager/main/SubjectDeleteExecuteAction.java
@@ -2,9 +2,11 @@ package scoremanager.main;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import jakarta.servlet.http.HttpSession;
 
 import bean.School;
 import bean.Subject;
+import bean.Teacher;
 import dao.SubjectDao;
 import tool.Action;
 
@@ -12,16 +14,23 @@ public class SubjectDeleteExecuteAction extends Action {
 	@Override
 	public void execute(HttpServletRequest request, HttpServletResponse response) throws Exception{
 		
-		School school = (School) request.getSession().getAttribute("school"); 
+		HttpSession session = request.getSession();
+		Teacher teacher = (Teacher)session.getAttribute("teacher");
+		if (teacher == null) {
+			response.sendRedirect("Login.action");
+			return;
+		}
+		
 		//科目コード所得
 		String cd = request.getParameter("cd");
 		
 		Subject subject = new Subject();
+		School school = teacher.getSchool();
 		subject.setCd(cd);
 		
 		SubjectDao dao =new SubjectDao();
 		dao.delete(school, subject);
 		
-		request.getRequestDispatcher("subject_delete_done.jsp").forward(request, response);
+		request.getRequestDispatcher("/scoremanager/main/subject_delete_done.jsp").forward(request, response);
 	}
 }
